Add .env.example showing what keys will unlock useful stuff
When I get around to adding actual things this can query might have to
put either into comments in the env.example or README.md what kind of
find grained permissions we're asking for in a token here.

Would be good to ensure it's easy to meet principle of least privilidge
in minimal tokens and expanding that when a new usecase pops up.