Move data loaders back into the data directory
Right I've discovered it was a mistake to move these. Components are
client side javascript that runs in the browser / obesrvable notebook to
do things at runtime.

Data loaders are build-time scripts.

I've ended up in this mess because I'm writing the JSON to files and not
outputting it in the correct way, so first let's move everything back
then next address how things should actually be output