Add analysis of how long PRs take to be reviewed
Adds a new data loader which pulls data from the Github API about when
PRs are created, reviewed and merged. This currently operates on a hard
coded list of repositories; if you add a private repo to that list then
you'll also need to add a Github access token with read permissions to
your `.env` file.

This includes some questionable but probably accurate enough
calculations of working hours between two datetimes. I think these are
ok for this sort of broad analysis but don't copy this logic to a
production system!

The data loader is pretty slow - it needs to make 1 request to the API
per pull request to get details of the reviews. It's ok when looking at
a couple of repos, but it might not work well if we wanted to look
across the whole org. I filter out PRs by dependabot as early as
possible to speed things up.