Add .node-version file to make explicit intended version
I think this is generally good practice.
Also let's bump the node version to 22.9.0 for good measure.

That's not a version we're generally using on our prod services yet, but
given the status of this repo I think it's OK to be reasonably cutting
edge about things

Finally while we're at it, let's bump the dev dependency for node types